+++ src/main/java/org/openlmis/referencedata/domain/FacilityTypeApprovedProduct.java
@@ -81,13 +81,14 @@ public class FacilityTypeApprovedProduct extends BaseEntity {
    * @param importer instance of {@link FacilityTypeApprovedProduct.Importer}
    * @return new instance of FacilityTypeApprovedProduct.
    */
-  public static FacilityTypeApprovedProduct newFacilityTypeApprovedProduct(Importer importer) {
+  public static FacilityTypeApprovedProduct newFacilityTypeApprovedProduct(Importer importer,
+                                                                           Orderable orderable) {
     FacilityTypeApprovedProduct ftap = new FacilityTypeApprovedProduct();
     ftap.setId(importer.getId());
     ftap.setMaxPeriodsOfStock(importer.getMaxPeriodsOfStock());
     ftap.setMinPeriodsOfStock(importer.getMinPeriodsOfStock());
     ftap.setEmergencyOrderPoint(importer.getEmergencyOrderPoint());
-    ftap.setOrderable(Orderable.newInstance(importer.getOrderable()));
+    ftap.setOrderable(orderable);
     if (null != importer.getProgram()) {
       ftap.setProgram(Program.newProgram(importer.getProgram()));
     }